Fix all (previously broken) tests.
The tests were broken because they were waiting for the "Screen #" navigation bar title to be gone, but the navigation bar would then contain a back button (in some cases) with that title. Now there is a separate label in the demo view so you can check for the appearance/absence of the navigation bar title and view separately.